DPL: move writers to plugin
This will allow us at some point to remove the dependency on TTree in the framework,
hopefully saving memory and allowing us to more easily customize the writing backend.
