+++ Framework/Core/test/Mocking.h
@@ -34,7 +34,10 @@ std::unique_ptr<o2::framework::ConfigContext> makeEmptyConfigContext()
   store->preload();
   store->activate();
   static ConfigParamRegistry registry(std::move(store));
-  auto context = std::make_unique<ConfigContext>(registry, 0, nullptr);
+  static std::unique_ptr<ServiceRegistry> services;
+  // We need to reset it because we will inject services into it.
+  services = std::make_unique<ServiceRegistry>();
+  auto context = std::make_unique<ConfigContext>(registry, ServiceRegistryRef{*services}, 0, nullptr);
   return context;
 }
